requirements.txt
spleeter
langchain
tqdm
moviepy
pytube
pydub
whisper_timestamped
validators
python-docx
transformers
python-dotenv
openai
pinecone-client
Chromadb
langdetect 
numba
llvmlite
numpy